arch/boot: optimize setup_sregs and page initialization
separate the page setup process in the setup_sregs to make clear
boundary between dbs-arch and dbs-boot.
