+++ crates/dbs-arch/Cargo.toml
@@ -10,7 +10,7 @@ keywords = ["dragonball", "secure-sandbox", "arch", "ARM64", "x86", "VMM"]
 readme = "README.md"
 
 [dependencies]
-kvm-bindings = { version = ">=0.5.0", features = ["fam-wrappers"] }
+kvm-bindings = { version = "0.5.0", features = ["fam-wrappers"] }
 kvm-ioctls = ">=0.9.0"
 vm-memory = { version = "0.7" }
 vmm-sys-util = "0.9.0"
