+++ crates/dbs-boot/Cargo.toml
@@ -11,6 +11,8 @@ readme = "README.md"
 
 [dependencies]
 dbs-arch = { path = "../dbs-arch", version = "0.1" }
+kvm-bindings = { version = "0.5.0", features = ["fam-wrappers"] }
+kvm-ioctls = ">=0.9.0"
 lazy_static = "1"
 libc = "0.2.39"
 thiserror = "1"
